+++ code/items/units/Tank.cs
@@ -48,9 +48,9 @@ namespace Facepunch.RTS.Units
 			"tech.armoredplating",
 			"tech.pyrotechnics"
 		};
-		public override Dictionary<string, int> Resistances => new()
+		public override Dictionary<string, float> Resistances => new()
 		{
-			["resistance.fire"] = 0.2
+			["resistance.fire"] = 0.2f
 		};
 	}
 }
